+++ app/package.json
@@ -1,52 +1,27 @@
 {
-  "name": "gratiotquest",
+  "name": "app",
   "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.1.1",
-    "@testing-library/user-event": "^14.1.1",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^17.0.30",
-    "@types/react": "^18.0.8",
-    "@types/react-dom": "^18.0.3",
-    "date-fns": "^2.28.0",
-    "next": "^12.1.5",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "typescript": "^4.6.4",
-    "web-vitals": "^2.1.4"
-  },
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "predeploy": "yarn build",
-    "deploy": "gh-pages -d build"
+    "lint": "next lint"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ],
-    "rules": {
-      "curly": "error"
-    }
+  "dependencies": {
+    "@typescript-eslint/eslint-plugin": "^7.1.1",
+    "@typescript-eslint/parser": "^7.1.1",
+    "next": "14.1.3",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "homepage": ".",
   "devDependencies": {
-    "gh-pages": "^3.2.3"
+    "@types/node": "^20.0.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "eslint": "^8.0.0",
+    "eslint-config-next": "^14.1.3",
+    "prettier": "^3.2.5",
+    "typescript": "^5.0.0"
   }
 }
