--- app/pages/_app.tsx
@@ -1,5 +1,0 @@
-import '../styles/index.css'
-
-export default function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
-}
