--- app/styles/index.css
@@ -1,64 +1,0 @@
-body {
-  margin: 0;
-}
-
-* {
-  box-sizing: border-box;
-}
-
-.chat {
-  padding: 3px;
-  position: fixed;
-  bottom: 0;
-  width: 100%;
-}
-.chat input {
-  padding: 10px;
-  width: 90%;
-  margin-right: 0.5%;
-}
-.chat button {
-  width: 9%;
-  background: rgb(130, 224, 255);
-  padding: 10px;
-}
-
-#messages {
-  overflow: auto;
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-}
-#messages li {
-  padding: 5px 10px;
-}
-#messages li:nth-child(odd) {
-  background: #eee;
-}
-
-.container {
-  display: flex;
-}
-
-@media only screen and (max-width: 1000px) {
-  /* For mobile phones: */
-  .container {
-    flex-direction: column;
-  }
-  .chat input {
-    height: 100px;
-    background-color: green;
-  }
-  .chat button {
-    height: 100px;
-  }
-}
-
-.error {
-  color: red;
-}
-
-.userdialog {
-  width: 800px;
-  height: 400px;
-}
